put use of unexported APIs behind a build tag
